Organize tests and remove bug terminology
